Move params to folder inside entity
