+++ app/src/main/java/com/udinesfata/expenz/domain/entity/params/BudgetParams.kt
@@ -1,4 +1,4 @@
-package com.udinesfata.expenz.domain.params
+package com.udinesfata.expenz.domain.entity.params
 
 class BudgetParams(
     private val isActive: Boolean? = null
