+++ app/src/main/java/com/udinesfata/expenz/domain/entity/params/CategoryParams.kt
@@ -1,0 +1,5 @@
+package com.udinesfata.expenz.domain.entity.params
+
+class CategoryParams(
+
+)
