+++ app/src/main/java/com/udinesfata/expenz/domain/entity/params/WalletParams.kt
@@ -1,0 +1,3 @@
+package com.udinesfata.expenz.domain.entity.params
+
+class WalletParams()
