--- app/src/main/java/com/udinesfata/expenz/domain/params/CategoryParams.kt
@@ -1,5 +1,0 @@
-package com.udinesfata.expenz.domain.params
-
-class CategoryParams(
-
-)
