--- app/src/main/java/com/udinesfata/expenz/domain/params/WalletParams.kt
@@ -1,3 +1,0 @@
-package com.udinesfata.expenz.domain.params
-
-class WalletParams()
