+++ app/src/main/java/com/udinesfata/expenz/domain/usecase/GetAllWalletsUseCase.kt
@@ -1,7 +1,7 @@
 package com.udinesfata.expenz.domain.usecase
 
 import com.udinesfata.expenz.domain.entity.Wallet
-import com.udinesfata.expenz.domain.params.WalletParams
+import com.udinesfata.expenz.domain.entity.params.WalletParams
 import com.udinesfata.expenz.domain.repository.WalletRepository
 
 class GetAllWalletsUseCase(
