+++ app/src/main/java/com/udinesfata/expenz/domain/usecase/GetTransactionsByWalletUseCase.kt
@@ -1,7 +1,7 @@
 package com.udinesfata.expenz.domain.usecase
 
 import com.udinesfata.expenz.domain.entity.Transaction
-import com.udinesfata.expenz.domain.params.TransactionParams
+import com.udinesfata.expenz.domain.entity.params.TransactionParams
 import com.udinesfata.expenz.domain.repository.TransactionRepository
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
