+++ app/src/main/java/com/udinesfata/expenz/domain/usecase/GetTransactionsTimelineUseCase.kt
@@ -1,8 +1,8 @@
 package com.udinesfata.expenz.domain.usecase
 
 import com.udinesfata.expenz.domain.entity.Transaction
-import com.udinesfata.expenz.domain.params.ORDER_DESC
-import com.udinesfata.expenz.domain.params.TransactionParams
+import com.udinesfata.expenz.domain.entity.params.ORDER_DESC
+import com.udinesfata.expenz.domain.entity.params.TransactionParams
 import com.udinesfata.expenz.domain.repository.TransactionRepository
 
 class GetTransactionsTimelineUseCase(
